Update MergeTeslaCam.py
A Tesla error resulted in a filename with no timestamp prefix. It was named "-event.json". A filename that starts with a hyphen breaks stuff because it looks like a command line switch.
This should be rewritten to be "only if good" instead of this "unless bad in this specific way" solution, but this change does trap and skip over any "-event.json" file so the program can continue.